+++ lib/perfect_shape/composite_shape.rb
@@ -63,10 +63,11 @@ module PerfectShape
     # @return true if the point lies within the bound of
     # the composite shape or false if the point lies outside of the
     # path's bounds.
-    def contain?(x_or_point, y = nil)
+    def contain?(x_or_point, y = nil, outline: false, distance_tolerance: 0)
       x, y = normalize_point(x_or_point, y)
       return unless x && y
-      shapes.any? {|shape| shape.contain?(x, y) }
+      
+      shapes.any? { |shape| shape.contain?(x, y, outline: outline, distance_tolerance: distance_tolerance) }
     end
   end
 end