+++ lib/perfect_shape/rectangle.rb
@@ -40,6 +40,7 @@ module PerfectShape
     def contain?(x_or_point, y = nil, outline: false, distance_tolerance: 0)
       x, y = normalize_point(x_or_point, y)
       return unless x && y
+      
       if outline
         edges.any? { |edge| edge.contain?(x, y, distance_tolerance: distance_tolerance) }
       else